feat(midi): implement DAW-specific MIDI note calibration
- Add persistent midiOctaveOffset to ConfigManager.
- Implement MIDI calibration mode in DuqAudioProcessor using AsyncUpdater.
- Refactor MidiUtils to resolve circular header dependencies.
- Add 'Listen for C0' feature to Workflow settings with real-time feedback.
- Synchronize envelope triggers and piano modal with calibrated offset.
- Fix various integer-to-float conversion warnings and layout overlapping.

+++ src/ui/components/WaveformComponent.cpp
@@ -142,6 +142,6 @@ void WaveformComponent::paint(juce::Graphics& g)
     float nyCenter = (0.5f - offsetY) / visibleHeightNorm;
     float yCenter = (1.0f - nyCenter) * height;
     g.setColour(T_COL(gridMinor));
-    g.drawLine(0, yCenter, width, yCenter);
+    g.drawLine(0.0f, yCenter, (float)width, yCenter);
 }
